requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
asgiref=3.2.10=py_0
ca-certificates=2020.7.22=0
certifi=2020.6.20=py38_0
django=3.0.3=py_0
django-filter=2.3.0=py_0
freetype=2.10.2=hd328e21_0
jpeg=9b=hb83a4c4_2
libpng=1.6.37=h2a8f88b_0
libtiff=4.1.0=h56a325e_1
lz4-c=1.9.2=h62dcd97_1
olefile=0.46=py_0
openssl=1.1.1g=he774522_1
pillow=7.2.0=py38hcc1f983_0
pip=20.2.2=py38_0
python=3.8.5=he1778fa_0
pytz=2020.1=py_0
setuptools=49.6.0=py38_0
sqlite=3.33.0=h2a8f88b_0
sqlparse=0.3.1=py_0
tk=8.6.10=he774522_0
vc=14.1=h0510ff6_4
vs2015_runtime=14.16.27012=hf0eaf9b_3
wheel=0.35.1=py_0
wincertstore=0.2=py38_0
xz=5.2.5=h62dcd97_0
zlib=1.2.11=h62dcd97_4
zstd=1.4.5=h04227a9_0
